fix(accounts): proper OAuth identity resolution on private emails
Before this patch, if an user authenticated through GitHub had a private email in the former platform, we would redirect them to our `/provide-email` endpoint as we need their email address to be able to communicate with them and let the user have a way to recover their account. However, if the user was required to re-authenticate they would be redirected back to `/provide-email` as the Accounts context wouldn't check for existing accounts but try to find the user through their email address provided by the OAuth service making them unable to re-authenticate.